+++ src/corpus/CorpusDelete.vue
@@ -6,17 +6,17 @@ import RouteButton from "@/components/RouteButton.vue";
 import LayoutSection from "@/components/LayoutSection.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 import useDeleteCorpus from "./deleteCorpus.composable";
-import { useCorpusStore } from "@/store/corpus.store";
+import { useResourceStore } from "@/store/resource.store";
 
 const router = useRouter();
 const corpusId = useCorpusIdParam();
 const { deleteCorpus } = useDeleteCorpus();
-const corpusStore = useCorpusStore();
+const resourceStore = useResourceStore();
 
 async function doDelete() {
   await deleteCorpus(corpusId);
-  if (!(corpusId in corpusStore.corpora)) {
-    router.push("/corpus");
+  if (!(corpusId in resourceStore.corpora)) {
+    router.push("/library");
   }
 }
 </script>
@@ -24,7 +24,7 @@ async function doDelete() {
 <template>
   <PendingContent :on="`corpus/${corpusId}`" blocking>
     <LayoutSection :title="$t('corpus.delete')">
-      <p class="my-2">{{ $t("corpus.delete.ask") }}</p>
+      <p class="my-2">{{ $t("resource.delete.ask") }}</p>
 
       <div>
         <ActionButton class="button-danger mr-4" @click="doDelete">
@@ -32,7 +32,7 @@ async function doDelete() {
           {{ $t("corpus.delete") }}
         </ActionButton>
 
-        <RouteButton :to="`/corpus/${corpusId}`">{{
+        <RouteButton :to="`/library/corpus/${corpusId}`">{{
           $t("cancel")
         }}</RouteButton>
       </div>