+++ src/corpus/config/config.composable.ts
@@ -5,23 +5,21 @@ import {
   parseConfig,
   type ConfigOptions,
 } from "@/api/corpusConfig";
-import useLocale from "@/i18n/locale.composable";
 import useMinkBackend from "@/api/backend.composable";
-import { useCorpusStore } from "@/store/corpus.store";
+import { useResourceStore } from "@/store/resource.store";
 
 export default function useConfig(corpusId: string) {
-  const corpusStore = useCorpusStore();
-  const { th } = useLocale();
+  const resourceStore = useResourceStore();
   const mink = useMinkBackend();
 
-  const corpus = computed(() => corpusStore.corpora[corpusId]);
+  const corpus = computed(() => resourceStore.corpora[corpusId]);
   const config = computed(() => corpus.value?.config);
-  const corpusName = computed(() => th(config.value?.name));
 
   async function loadConfig() {
     const config = await mink
       .loadConfig(corpusId)
       .then(parseConfig)
+      // 404 means no config which is fine, rethrow other errors.
       .catch((error) => {
         if (error.response?.status == 404) return emptyConfig();
         throw error;
@@ -32,12 +30,11 @@ export default function useConfig(corpusId: string) {
   async function uploadConfig(config: ConfigOptions) {
     // This may throw, either from makeConfig or saveConfig.
     await mink.saveConfig(corpusId, await makeConfig(corpusId, config));
-    corpusStore.corpora[corpusId].config = config;
+    resourceStore.corpora[corpusId].config = config;
   }
 
   return {
     config,
-    corpusName,
     loadConfig,
     uploadConfig,
   };