+++ src/corpus/sources/SourceText.vue
@@ -2,6 +2,7 @@
 import { onMounted, ref } from "vue";
 import { downloadFile } from "@/util";
 import ActionButton from "@/components/ActionButton.vue";
+import TextData from "@/components/TextData.vue";
 
 const props = defineProps<{
   load: () => Promise<string | undefined>;
@@ -10,7 +11,6 @@ const props = defineProps<{
 }>();
 
 const text = ref();
-const expanded = ref(false);
 const loadPromise = props.load();
 
 onMounted(async () => {
@@ -21,51 +21,15 @@ async function download() {
   const text = await loadPromise;
   downloadFile(text!, props.filename || "mink-source");
 }
-
-function toggleExpand() {
-  expanded.value = !expanded.value;
-}
 </script>
 
 <template>
-  <div
-    v-if="text"
-    class="relative bg-zinc-700 dark:bg-zinc-600 rounded shadow-inner text-sm"
-  >
-    <div class="absolute z-10 bottom-2 right-2 flex gap-2">
-      <ActionButton @click="toggleExpand">
-        <template v-if="expanded">
-          <icon :icon="['far', 'square-minus']" class="mr-1" />
-          {{ $t("expand.close") }}
-        </template>
-        <template v-else>
-          <icon :icon="['far', 'square-plus']" class="mr-1" />
+  <TextData v-if="text" :text="text" />
 
-          {{ $t("expand.open") }}
-        </template>
-      </ActionButton>
-      <ActionButton @click="download">
-        <icon :icon="['far', 'file']" class="mr-1" />
-        {{ $t("download") }}
-      </ActionButton>
-    </div>
-    <div
-      class="text-white overflow-hidden whitespace-pre-wrap font-mono text-xs p-2 pb-10"
-      :class="{ 'h-20': !expanded }"
-    >
-      {{ expanded ? text : text.slice(0, 800) }}
-    </div>
-  </div>
-  <div v-else>
+  <div class="my-2">
     <ActionButton @click="download">
       <icon :icon="['far', 'file']" class="mr-1" />
       {{ $t("download") }}
     </ActionButton>
   </div>
 </template>
-
-<style scoped>
-.h-20 {
-  mask-image: linear-gradient(black 50%, transparent);
-}
-</style>