+++ src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 import App from "./App.vue";
-import router from "./router";
+import router from "./router/router";
 import { createPinia } from "pinia";
 import i18n from "./i18n/i18n";
 import { formkit, formkitConfig } from "./formkit";