+++ src/resource/resourceIdParam.composable.ts
@@ -1,0 +1,6 @@
+import { useRoute } from "vue-router";
+
+export default function useResourceIdParam() {
+  const route = useRoute();
+  return route.params.resourceId as string;
+}